Chrome_Extension/quiz_generator_2/requirements.txt
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain_google_genai>=1.0.0
python-dotenv